// src/types/inspection.ts
export interface InspectionProps {
  id: number;
  regenerator: string;
  inspector: string;
  status: number;
  acceptedAt: number;
  createdAt: number;
  inspectedAt: number;
  inspectedAtEra: number;
  invalidatedAt: number;
  proofPhoto: string;
  regenerationScore: number;
  report: string;
  treesResult: number;
  validationsCount: number;
}